refactor: Http의 request와 response로 패키지 분리
