+++ src/main/java/http/HttpVersion.java
@@ -1,4 +1,4 @@
-package webserver;
+package http;
 
 import java.util.Arrays;
