+++ src/main/java/http/request/RequestUri.java
@@ -1,29 +1,29 @@
-package webserver;
+package http.request;
 
 import java.util.Objects;
 
-public class HttpRequestUri {
+public class RequestUri {
     private static final String DELIMITER = "\\?";
 
     private final String path;
     private final QueryParameters queryParameters;
 
-    private HttpRequestUri(String path, QueryParameters queryParameters) {
+    private RequestUri(String path, QueryParameters queryParameters) {
         this.path = path;
         this.queryParameters = queryParameters;
     }
 
-    private HttpRequestUri(String path) {
+    private RequestUri(String path) {
         this(path, null);
     }
 
-    public static HttpRequestUri from(String requestUri) {
+    public static RequestUri from(String requestUri) {
         Objects.requireNonNull(requestUri);
         if (hasQueryParameters(requestUri)) {
-            return new HttpRequestUri(requestUri);
+            return new RequestUri(requestUri);
         }
         String[] tokens = requestUri.split(DELIMITER);
-        return new HttpRequestUri(tokens[0], QueryParameters.from(tokens[1]));
+        return new RequestUri(tokens[0], QueryParameters.from(tokens[1]));
     }
 
     private static boolean hasQueryParameters(String requestUri) {
